requirements.txt
scrapy==2.11.0
selenium==4.18.1
playwright==1.42.0
beautifulsoup4==4.12.2
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
nltk==3.8.1
spacy==3.7.4
hubspot-api-client==7.5.0
feedparser==6.0.11
python-dotenv==1.0.1
pytest==8.0.0
schedule==1.2.1
pydantic==2.6.1
sqlalchemy==2.0.23
tenacity==8.2.3
alembic==1.12.1
email-validator==2.1.0